Fix schematics with a high number of tiles not loading There is a maximum number of open files in php, so we can't clone the file handler for every tag This also fixes a lot of places where file handles weren't closed properly (due to the whole cloning magic)
